Restructure populate scripts
