+++ populate/src/bets/findCheaters.js
@@ -1,7 +1,7 @@
 /*
  * Check if someone bet after the beginning of a game
  */
-const { serviceAccount } = require('./chooseDatabase.js')
+const { serviceAccount } = require('../chooseDatabase.js')
 
 const admin = require('firebase-admin')
 admin.initializeApp({
