+++ populate/src/bets/whoDontBet.js
@@ -1,7 +1,7 @@
 /*
  * Display which user didn't bet on which match
  */
-const { serviceAccount } = require('./chooseDatabase.js')
+const { serviceAccount } = require('../chooseDatabase.js')
 
 const admin = require('firebase-admin')
 admin.initializeApp({
