+++ populate/src/database/addUserAdmin.js
@@ -1,4 +1,4 @@
-const { serviceAccount } = require('./chooseDatabase.js')
+const { serviceAccount } = require('../chooseDatabase.js')
 
 const admin = require('firebase-admin')
 admin.initializeApp({
