+++ populate/src/database/firestore-data/euro/stadiums.json
