+++ populate/src/notifications/send-notifications-matchs-update.js
@@ -1,7 +1,7 @@
 /**
  * Script to test notifications in app.
  */
-const { serviceAccount } = require('./chooseDatabase.js')
+const { serviceAccount } = require('../chooseDatabase.js')
 
 const admin = require('firebase-admin')
 admin.initializeApp({
