+++ populate/src/old_scripts/rankingGlobalRtR.js
