+++ populate/src/ranking/rankingGlobal.js
@@ -1,7 +1,7 @@
 /*
  * Display tribes of all opponents
  */
-const { serviceAccount } = require('./chooseDatabase.js')
+const { serviceAccount } = require('../chooseDatabase.js')
 
 const admin = require('firebase-admin')
 admin.initializeApp({
