+++ populate/src/tribes/whoHaveNoTribe.js
@@ -1,7 +1,7 @@
 /*
  * Display user with no tribe
  */
-const { serviceAccount } = require('./chooseDatabase.js')
+const { serviceAccount } = require('../chooseDatabase.js')
 
 const admin = require('firebase-admin')
 admin.initializeApp({
